Refactor console control; move printContext and printCommands functions to console_control.cpp, update usage in ProcessController, and improve command help output
